requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.10.1
pypdf==3.17.1
python-dotenv==1.0.1
bcrypt==4.2.0
